requirements.txt
streamlit
numpy
pandas
seaborn
matplotlib
plotly
scikit_learn
pillow